(#39) settings: extract into a new module, visible for both stack and main
