+++ settings/src/main/scala/me/fornever/haskeletor/settings/HaskellSettingsState.scala
