+++ src/main/scala/me/fornever/haskeletor/action/OrmoluReformatAction.scala
@@ -13,10 +13,10 @@ import com.intellij.openapi.actionSystem.{AnAction, AnActionEvent}
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
-import me.fornever.haskeletor.HTool
 import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.external.component.StackProjectManager
 import me.fornever.haskeletor.external.execution.CommandLine
+import me.fornever.haskeletor.settings.HTool
 import me.fornever.haskeletor.util._
 
 class OrmoluReformatAction extends AnAction {
