+++ src/main/scala/me/fornever/haskeletor/external/execution/CommandLine.scala
@@ -17,9 +17,9 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.util.io.BaseOutputReader
-import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
 import me.fornever.haskeletor.core.stack.OutputToProgressIndicator
+import me.fornever.haskeletor.settings.GlobalInfo
 import org.jetbrains.annotations.Nls
 
 import java.nio.charset.Charset
