+++ src/main/scala/me/fornever/haskeletor/external/repl/StackRepl.scala
@@ -9,14 +9,14 @@
 package me.fornever.haskeletor.external.repl
 
 import com.intellij.openapi.project.Project
-import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
-import me.fornever.haskeletor.core.util.StringUtil
+import me.fornever.haskeletor.core.util.{FileSystemUtil, StringUtil}
 import me.fornever.haskeletor.external.execution.StackCommandLine
 import me.fornever.haskeletor.external.repl.StackRepl.{BenchmarkType, ExeType, StackReplOutput, TestSuiteType}
 import me.fornever.haskeletor.external.repl.StackReplsManager.ProjectReplTargets
 import me.fornever.haskeletor.sdk.HaskellSdkType
-import me.fornever.haskeletor.util.{HaskellEditorUtil, HaskellFileUtil, HaskellProjectUtil}
+import me.fornever.haskeletor.settings.GlobalInfo
+import me.fornever.haskeletor.util.{HaskellEditorUtil, HaskellProjectUtil}
 
 import java.io._
 import java.util.concurrent.LinkedBlockingQueue
@@ -317,7 +317,7 @@ abstract class StackRepl(project: Project, projectReplTargets: Option[ProjectRep
     if (!ghciOptionsFile.exists()) {
       ghciOptionsFile.createNewFile()
       ghciOptionsFile.setWritable(true, true)
-      HaskellFileUtil.removeGroupWritePermission(ghciOptionsFile)
+      FileSystemUtil.removeGroupWritePermission(ghciOptionsFile)
 
       val writer = new BufferedWriter(new FileWriter(ghciOptionsFile))
       try {
