+++ src/main/scala/me/fornever/haskeletor/module/HaskellModuleBuilder.scala
@@ -26,7 +26,6 @@ import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.platform.templates.TemplateModuleBuilder
 import com.intellij.util.ui.JBUI
-import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.cabal.PackageInfo
 import me.fornever.haskeletor.core.HaskeletorBundle
 import me.fornever.haskeletor.core.notifications.HaskellNotificationGroup
@@ -34,7 +33,7 @@ import me.fornever.haskeletor.external.component.{HaskellComponentsManager, Libr
 import me.fornever.haskeletor.external.execution.{CommandLine, StackCommandLine}
 import me.fornever.haskeletor.icons.HaskellIcons
 import me.fornever.haskeletor.sdk.HaskellSdkType
-import me.fornever.haskeletor.settings.HaskellSettingsState
+import me.fornever.haskeletor.settings.{GlobalInfo, HaskellSettingsState}
 import me.fornever.haskeletor.stackyaml.StackYamlComponent
 import me.fornever.haskeletor.util.{FutureUtil, HaskellFileUtil, HaskellProjectUtil, ScalaUtil}
