+++ src/main/scala/me/fornever/haskeletor/runconfig/console/HaskellConsoleState.scala
@@ -14,10 +14,11 @@ import com.intellij.execution.filters.TextConsoleBuilderImpl
 import com.intellij.execution.process.{ProcessHandler, ProcessTerminatedListener}
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.ui.ConsoleView
-import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.core.project.GhcVersion
+import me.fornever.haskeletor.core.util.FileSystemUtil
 import me.fornever.haskeletor.external.component.HaskellComponentsManager
 import me.fornever.haskeletor.sdk.HaskellSdkType
+import me.fornever.haskeletor.settings.GlobalInfo
 import me.fornever.haskeletor.util.HaskellFileUtil
 
 import java.io.File
@@ -45,7 +46,7 @@ class HaskellConsoleState(val configuration: HaskellConsoleConfiguration, val en
         if (!ghciScript.exists()) {
           HaskellFileUtil.copyStreamToFile(getClass.getResourceAsStream(s"/ghci/$ghciScriptName"), ghciScript)
           ghciScript.setWritable(true, true)
-          HaskellFileUtil.removeGroupWritePermission(ghciScript)
+          FileSystemUtil.removeGroupWritePermission(ghciScript)
         }
 
         val commandLine = new GeneralCommandLine(stackPath)
