+++ src/main/scala/me/fornever/haskeletor/util/HaskellProjectUtil.scala
@@ -16,11 +16,11 @@ import com.intellij.openapi.vfs.{VfsUtilCore, VirtualFile}
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.{PsiElement, PsiFile}
 import com.intellij.util.PathUtilRt
-import me.fornever.haskeletor.GlobalInfo
 import me.fornever.haskeletor.core.project.GhcVersion
 import me.fornever.haskeletor.external.component.HaskellComponentsManager
 import me.fornever.haskeletor.module.HaskellModuleType
 import me.fornever.haskeletor.sdk.HaskellSdkType
+import me.fornever.haskeletor.settings.GlobalInfo
 import org.jetbrains.jps.model.java.JavaModuleSourceRootTypes
 
 import java.io.File
